% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antideg_poly.r
\docType{data}
\name{antideg_poly}
\alias{antideg_poly}
\title{Utah antidegradation category polyon}
\format{
An sf type polygon shapefile
}
\usage{
antideg_poly
}
\description{
Polygons containing antidegradation category delineations.
}
\keyword{datasets}
