% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignAUs.r
\name{assignAUs}
\alias{assignAUs}
\title{Assign Utah assessment units to sites}
\usage{
assignAUs(data, lat = "LatitudeMeasure", long = "LongitudeMeasure")
}
\arguments{
\item{data}{Input dataset. Must include latitude & longitude columns.}

\item{lat}{Name of latitude column. Default matches WQP objects.}

\item{long}{Name of longitude column. Default matches WQP objects.}
}
\value{
Returns the input data frame with assessment unit information appended.
}
\description{
This function assigns assessment units to water quality portal type site objects (or data with site information attached). This can be done before or after assigning beneficial uses.
}
\examples{
# Read a couple of sites (Mantua Reservoir and a tributary)
sites=readWQP(type="sites", siteid=c("UTAHDWQ_WQX-4900440","UTAHDWQ_WQX-4900470"))
sites_AUs=assignAUs(sites)
}
