% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignBasin.r
\name{assignBasin}
\alias{assignBasin}
\title{Assign basin (management unit) to records}
\usage{
assignBasin(
  data,
  lat = "LatitudeMeasure",
  long = "LongitudeMeasure",
  assess_id = "ASSESS_ID",
  basin_col = "Mgmt_Unit"
)
}
\arguments{
\item{data}{Input dataset.}

\item{lat}{Name of latitude column. Default matches WQP objects.}

\item{long}{Name of longitude column. Default matches WQP objects.}

\item{assess_id}{Name of assessment unit id column used for fallback.
Defaults to \code{ASSESS_ID}.}

\item{basin_col}{Name of basin column in \code{wmu_poly} to return.
Defaults to \code{Mgmt_Unit}.}
}
\value{
Returns the input data frame with basin information appended.
}
\description{
This function assigns a basin field (default \code{Mgmt_Unit}) using
\code{wqTools::wmu_poly}. It uses latitude/longitude when available and then
fills any remaining missing basin values using \code{ASSESS_ID}.
}
\examples{
# Read a couple of sites and assign basin
sites=readWQP(type="sites", siteid=c("UTAHDWQ_WQX-4900440","UTAHDWQ_WQX-4900470"))
sites_basin=assignBasin(sites)
}
