% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPolys.r
\name{assignPolys}
\alias{assignPolys}
\title{Assign spatial polygons to sites}
\usage{
assignPolys(
  data,
  polygon,
  lat = "LatitudeMeasure",
  long = "LongitudeMeasure",
  columns
)
}
\arguments{
\item{data}{Input dataset. Must include latitude & longitude columns.}

\item{polygon}{Polygon to assign information to sites. Must be an sf type polygon shapefile, WGS84 projection recommended.}

\item{lat}{Name of latitude column. Default matches WQP objects.}

\item{long}{Name of longitude column. Default matches WQP objects.}

\item{columns}{Vector of columns to retain from polygon object. If columns not specified, function will retain all columns.}
}
\value{
Returns the input data frame with assessment unit information appended.
}
\description{
This function assigns spatial polygon data to water quality portal type site objects (or data with site information attached).
}
\examples{
# Read a couple of sites (Mantua Reservoir and a tributary)
sites=readWQP(type="sites", siteid=c("UTAHDWQ_WQX-4900440","UTAHDWQ_WQX-4900470"))
sites_AUs=assignPolys(sites, wqTools::au_poly)
sites_BUs=assignPolys(sites, wqTools::bu_poly)
}
