% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignUses.r
\name{assignUses}
\alias{assignUses}
\title{Assign Utah beneficial use classes to sites}
\usage{
assignUses(
  data,
  lat = "LatitudeMeasure",
  long = "LongitudeMeasure",
  flatten = FALSE
)
}
\arguments{
\item{data}{Input dataset. Must include latitude & longitude columns.}

\item{lat}{Name of latitude column. Default matches WQP objects.}

\item{long}{Name of longitude column. Default matches WQP objects.}

\item{flatten}{Logical. If FALSE (default), maintain use categorys as single comma separated column. If TRUE, use column and data are flattened by expanded use column.}
}
\description{
This function assigns beneficial use classes to water quality portal type site objects (or data with site information attached).
}
\examples{
# Read a couple of sites (Mantua Reservoir and a tributary)
sites=readWQP(type="sites", siteid=c("UTAHDWQ_WQX-4900440","UTAHDWQ_WQX-4900470"))
sites_uses=assignUses(sites)
sites_uses_flat=assignUses(sites, flatten=TRUE)
}
