% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_poly.r
\docType{data}
\name{au_poly}
\alias{au_poly}
\title{Utah's IR-specific assessment unit polygons}
\format{An sf type polygon shapefile}
\usage{
au_poly
}
\description{
Polygons containing assessment unit designations.
}
\keyword{datasets}
