% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_poly_1820.r
\docType{data}
\name{au_poly_1820}
\alias{au_poly_1820}
\title{IR-specific assessment unit polygons as used in the 2018/2020 IR}
\format{
An sf type polygon shapefile
}
\usage{
au_poly_1820
}
\description{
Polygons containing assessment unit designations as used in the 2018/2020 IR.
}
\keyword{datasets}
