% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_poly_2022.r
\docType{data}
\name{au_poly_2022}
\alias{au_poly_2022}
\title{IR-specific assessment unit polygons as used in the 2022 IR}
\format{
An sf type polygon shapefile
}
\usage{
au_poly_2022
}
\description{
Polygons containing assessment unit designations as used in the 2022 IR.
}
\keyword{datasets}
