% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_poly_2024.r
\docType{data}
\name{au_poly_2024}
\alias{au_poly_2024}
\title{IR-specific assessment unit polygons as used in the 2024 IR}
\format{
An sf type polygon shapefile
}
\usage{
au_poly_2024
}
\description{
Polygons containing assessment unit designations as used in the 2024 IR.
}
\keyword{datasets}
