% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMap.R
\name{baseMap}
\alias{baseMap}
\title{Water quality base map}
\usage{
baseMap(layers_control = TRUE)
}
\arguments{
\item{layers_control}{Logical. If TRUE (default), a layers control widget is added to the map. Set to FALSE to generate a custom layers control widget.}
}
\value{
A leaflet map object
}
\description{
This function generates a base map for use in water quality analyses. It includes useful base layers, Utah assessment unit polygons, and Utah beneficial use designations.
}
