% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bu_poly.r
\docType{data}
\name{bu_poly}
\alias{bu_poly}
\title{Utah's beneficial use polygon shapes}
\format{
An sf type polygon shapefile
}
\usage{
bu_poly
}
\description{
Polygons containing beneficial use designations and water body type information. Used to assign uses or standards to site locations.
}
\keyword{datasets}
