% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTSI.r
\name{calcTSI}
\alias{calcTSI}
\title{Calculate TSI values from input data}
\usage{
calcTSI(
  x,
  in_format = "wide",
  chl_ugL = "Chlorophyll a",
  TP_mgL = "Phosphate-phosphorus",
  SD_m = "Depth, Secchi disk depth",
  value_var = "ResultMeasureValue",
  param_var = "CharacteristicName"
)
}
\arguments{
\item{x}{Input dataset}

\item{in_format}{One of "wide" or "flat" to specify data input format. Note that only wide format inputs are currently supported.}

\item{chl_ugL}{Name of chlorophyll-a variable in ug/L}

\item{TP_mgL}{Name of total phosphorus variable in mg/L}

\item{SD_m}{Name of secchi disk depth variable in m}
}
\description{
This function calculates TSI values according to Utah's IR methods from input data containing values for of chlorophyll, total phosphorus, and secchi disk depth.
Note that inputs for these parameters must be specified in units of ug/L, mg/L, and meters, respectively.
}
\examples{
data(ul_trophic)
head(ul_trophic)
tsi=calcTSI(ul_trophic,chl_ugL="ChlA",TP_mgL="Phosphate.phosphorus.Total",SD_m="Depth.Secchi.disk.depth")
head(tsi)
plot(TSIchl~ChlA,tsi)
}
