% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_poly.r
\docType{data}
\name{county_poly}
\alias{county_poly}
\title{Utah's Counties}
\format{
An sf type polygon shapefile
}
\usage{
county_poly
}
\description{
Polygons containing county boundaries from gis.utah.gov.
}
\keyword{datasets}
