% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facToNum.r
\name{facToNum}
\alias{facToNum}
\title{Convert factors to numeric equivalents
Converts input object to number if class=="factor". If class !="factor", input object is returned un-altered.}
\usage{
facToNum(x)
}
\arguments{
\item{x}{Input vector object}
}
\description{
Convert factors to numeric equivalents
Converts input object to number if class=="factor". If class !="factor", input object is returned un-altered.
}
