% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSites.r
\name{findSites}
\alias{findSites}
\title{Interactively find and select water quality sites}
\usage{
findSites()
}
\value{
A dataframe of site information
}
\description{
A Shiny gadget designed for interactive selection of water quality sites including permit locations, ambient monitoring locations, and streamflow gauges. This is intended to help users find sites for querying data from ECHO, WQP, or NWIS.
}
