% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnnc_poly.r
\docType{data}
\name{hnnc_poly}
\alias{hnnc_poly}
\title{Utah's headwater numeric nutrient criteria polygons}
\format{
An sf type polygon shapefile
}
\usage{
hnnc_poly
}
\description{
A spatial polygons representation of Utah's headwater numeric nutrient criteria.
}
\keyword{datasets}
