% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huc12_poly.r
\docType{data}
\name{huc12_poly}
\alias{huc12_poly}
\title{HUC 12 polygons for the state of Utah}
\format{
An sf type polygon shapefile
}
\usage{
huc12_poly
}
\description{
HUC 12 polygons for the state of Utah.
}
\keyword{datasets}
