% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huc8_12_poly.r
\docType{data}
\name{huc8_12_poly}
\alias{huc8_12_poly}
\title{Unioned HUC 8 & 12 polygons for the state of Utah}
\format{An sf type polygon shapefile}
\usage{
huc8_12_poly
}
\description{
Unioned HUC 8 & 12 polygons for the state of Utah.
}
\keyword{datasets}
