% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huc8_poly.r
\docType{data}
\name{huc8_poly}
\alias{huc8_poly}
\title{HUC 8 polygons for the state of Utah}
\format{
An sf type polygon shapefile
}
\usage{
huc8_poly
}
\description{
HUC 8 polygons for the state of Utah.
}
\keyword{datasets}
