% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni_poly.r
\docType{data}
\name{muni_poly}
\alias{muni_poly}
\title{Utah's Municipalities}
\format{
An sf type polygon shapefile
}
\usage{
muni_poly
}
\description{
Polygons containing municipal boundaries from gis.utah.gov.
}
\keyword{datasets}
