% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileHeatMap.r
\name{profileHeatMap}
\alias{profileHeatMap}
\title{Profile heat map plot}
\usage{
profileHeatMap(
  data,
  parameter,
  depth = "Depth_m",
  param_units,
  depth_units = "m",
  date = "ActivityStartDate",
  show_dates = TRUE,
  min_date = min(data[, date], na.rm = T),
  max_date = max(data[, date], na.rm = T),
  param_lab = "pH",
  criteria
)
}
\arguments{
\item{data}{Lake profile data (wide format)}

\item{parameter}{Column name for parameter to be used as z-values.}

\item{depth}{Column name for depth column.}

\item{param_units}{Character. Parameter units. Used to plot build label.}

\item{depth_units}{Character. Depth units. Used to plot build label.}

\item{date}{Date column name. Must be in 'YYYY-mm-dd' format.}

\item{show_dates}{Logical. If TRUE (default), show individual profile dates on plot x-axis.}

\item{min_date}{Minimum plot date. 'YYYY-mm-dd' format.}

\item{max_date}{Maximum plot date. 'YYYY-mm-dd' format.}

\item{param_lab}{Character. Label to be used for parameter name. Used to build plot label.}

\item{criteria}{Vector of criteria values to be used as contours on heatmap plot. If not specified, contours from 0-30 at increments of 5 are drawn.}
}
\description{
Plots a lake profile heatmap for a single site and parameter.
}
