% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLakeProfile.r
\name{profilePlot}
\alias{profilePlot}
\title{Plot an individual lake profile}
\usage{
profilePlot(data, parameter = "CharacteristicName",
  units = "ResultMeasure.MeasureUnitCode",
  depth = "Depth, data-logger (ported)", do = "Dissolved oxygen (DO)",
  temp = "Temperature, water", pH = "pH",
  value_var = "ResultMeasureValue", line_no = "DataLoggerLine",
  do_crit, temp_crit, pH_crit)
}
\arguments{
\item{data}{Lake profile data (long format)}

\item{parameter}{Column name containing parameter names.}

\item{units}{Column name containing data units.}

\item{depth}{Name of depth variable in input data.}

\item{do}{Name of dissolved oxygen variable in input data.}

\item{temp}{Name of temperature variable in input data.}

\item{pH}{Name of pH variable in input data.}

\item{value_var}{Column name of value variable.}

\item{line_no}{Column name containing line number.}

\item{do_crit}{Optional. Dissolved oxygen criterion to display on plot.}

\item{temp_crit}{Optional. Temperature criterion to display on plot.}

\item{pH_crit}{Optional. Vector of two pH criteria to display on plot.}
}
\description{
Plots an individual lake profile provided in long data format. If provided, dashed lines representing water quality criteria are also plotted.
Default arguments are set to take data from the water quality portal, however, they may be updated as needed to reflect different data sources.
}
\examples{
# Read in some profile data
nr=readWQP(type="narrowresult", siteid="UTAHDWQ_WQX-4938550", print=F)
act=readWQP(type="activity", siteid="UTAHDWQ_WQX-4938550", print=F)
nr_act=merge(nr, act, all.x=T)
profiles=nr_act[!is.na(nr_act$DataLoggerLine),] #Subset to profile data
table(droplevels(profiles$ActivityIdentifier)) #Find activity IDs associated w/ profiles
profilePlot(subset(profiles,ActivityIdentifier=="UTAHDWQ_WQX-BORFG051909-4938550-0519-Pr-F"))
profilePlot(subset(profiles,ActivityIdentifier=="UTAHDWQ_WQX-LC082807-210255-PR3855083007"), do_crit=4, temp_crit=20, pH_crit=c(6.5,9))
}
