% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readECHO_ec.r
\name{readECHO_ec}
\alias{readECHO_ec}
\title{Read effluent chart data from EPA ECHO webservices}
\usage{
readECHO_ec(
  ...,
  print = TRUE,
  stringsAsFactors = default.stringsAsFactors(),
  progress = TRUE
)
}
\arguments{
\item{...}{additional arguments to be passed to ECHO query path. See c optional arguments for effluent chart data reads.}

\item{print}{Logical. If TRUE (default), print summary table of facilities & parameters returned.}

\item{stringsAsFactors}{Logical. Passed to read.csv. See ?read.csv for more information.}

\item{p_id}{Permitted facility ID. Either a single text value (in quotes) or a vector of text values.}

\item{parameter_code}{Parameter code. Either a single text value (in quotes) or a vector of text values.}

\item{start_date}{Query start date in "mm/dd/yyyy" format.}

\item{end_date}{Query end date in "mm/dd/yyyy" format.}
}
\value{
A flat data frame of EPA ECHO effluent chart data
}
\description{
This function extracts effluent chart data from EPA ECHO for multiple stations & combinations of parameters. All arguments are optional except p_id. At least one p_id must be specified.
}
\examples{
#Extract effluent chart data for facility UT0025241, all outfalls
UT0025241_ec=readECHO_ec(p_id="UT0025241", start_date="01/01/2010", end_date="01/15/2019")
head(UT0025241_ec)

# Extract effluent total phosphorus data from outfall 001 for facility UT0025241 
UT0025241_tp_001=readECHO_ec(p_id="UT0025241", parameter_code="00665", outfall="001")
UT0025241_tp_001_effluent=UT0025241_tp_001[UT0025241_tp_001$monitoring_location_desc=="Effluent Gross",]
head(UT0025241_tp_001_effluent)

# Extract flow through facility from UT0021717
UT0021717_flow=readECHO_ec(p_id="UT0021717", parameter_code="50050")

# Extract flow & TP from UT0025241 & UT0021717
tp_flow=readECHO_ec(p_id=c("UT0025241","UT0021717"), parameter_code=c("50050","00665"))
}
