% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readECHO_fac.r
\name{readECHO_fac}
\alias{readECHO_fac}
\title{Read facility information from EPA ECHO webservices}
\usage{
readECHO_fac(..., df = FALSE)
}
\arguments{
\item{...}{Additional arguments to be passed to ECHO query path. See https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_cwa_rest_services_get_facility_info for optional arguments for facilities. Note that arguments for output are ignored.}

\item{df}{Logical. If TRUE, extract geometry coordinates from list format and return as columns in data frame.}
}
\value{
A data frame of EPA ECHO facility information
}
\description{
This function extracts facility information from EPA ECHO based on argument inputs.
}
\examples{
# Read facility locations in Utah
ut_fac=readECHO_fac(p_st="ut", p_act="y")
head(ut_fac)
# Read facility locations for two permit IDs
two_fac=readECHO_fac(p_pid=c("UT0021717","UT0025241"))
two_fac
}
