% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readULgageheight.R
\name{readULgageheight}
\alias{readULgageheight}
\title{Read daily Utah Lake gage heights}
\usage{
readULgageheight(cap_curve = "lidar", cap_curve_only = FALSE)
}
\arguments{
\item{cap_curve}{One of 'lidar' (default) or 'udwr' to select the capacity curve used for surface area and volume estimates. Capacity curves are stored
within the function as dput dataframes. These can be exported and used outside of this function via the cap_curve_only argument (see below).}

\item{cap_curve_only}{Logical. If TRUE, returns only the selected capacity curve. Default is FALSE.}
}
\value{
Returns a dataframe of daily gage heights, lake elevations, and surface area and volume estimates.
}
\description{
This function reads Utah Lake gage heights and interprets them to elevations and surface area and volume estimates. Surface area
and volume estimates are based on curves developed by James Martin and Scott Daly. Gage heights are read from Utah Division of
Water Rights https://www.waterrights.utah.gov/cgi-bin/dvrtview.exe?STATION_ID=9019&RECORD_YEAR=2020&Modinfo=Daily_Comma&Startup.
}
\examples{
ul_elev=readULgageheight()
head(ul_elev)
cap_curve_lidar=readULgageheight(cap_curve_only=T)
head(cap_curve_lidar)
}
