% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWQPbySite.r
\name{readWQPbySite}
\alias{readWQPbySite}
\title{Read WQP data by selecting sites in an interactive map}
\usage{
readWQPbySite(sites, map, mode = "click", types = c("narrowresult",
  "activity", "sites"), merge = T, sitetypes = c("Canal Drainage",
  "Canal Irrigation", "Canal Transport", "Lake",
  "Lake, Reservoir, Impoundment", "Reservoir", "River/Stream",
  "River/Stream Intermittent", "River/Stream Perennial", "Seep", "Spring",
  "Stream", "Stream: Canal", "Stream: Ditch", "Wetland",
  "Wetland Undifferentiated"), ...)
}
\arguments{
\item{sites}{Optional. A sites object containing lat/long columns named "LatitudeMeasure" & "LongitudeMeasure". If no sites object is provided, a Utah state-wide query of WQP will be generated for you.}

\item{map}{Optional. A map object to use as a background for site selection. If no map object is provided, a basic map will be generated for you.}

\item{mode}{Mode for map selection. One of "click" or "draw". Click allows site selection by clicking on individual sites. Draw allows site selection by polygons.}

\item{types}{Vector of data types to read from WQP for selected sites. See ?wqTools::readWQP for options.}

\item{merge}{Logical. If TRUE (default), merge all selected data types to single data frame. Merges are performed as left joins in order they are provided. If FALSE, return list of individual data objects.}

\item{sitetypes}{Vector of site types to be included in query (only used if sites argument is not provided).}

\item{...}{Other arguments to be passed to readWQP (e.g. start_date, end_date, characteristicName, etc.). See ?wqTools::readWQP for more info.}
}
\description{
This function allows the user to read WQP data by selecting desired sites in an interactive map and specifiying desired types of data and output.
}
\examples{
wqp_data=readWQPbySite(start_date="01/01/2016", end_date="12/31/2018")
}
