% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPolys.r
\name{selectPolys}
\alias{selectPolys}
\title{Interactively select spatial polygons to query or filter data}
\usage{
selectPolys(type = "au")
}
\arguments{
\item{type}{One of 'au', 'huc8', or 'huc12' to select assessment unit, HUC 8, or HUC 12 polygons}
}
\value{
A vector of selected polygon identifiers (AU IDs or HUC numbers).
}
\description{
A Shiny gadget designed for interactive selection of spatial polygons. This is intended to help users find AU and HUC IDs for querying or filtering data.
}
\examples{
sites=readWQP(type='sites', auid=selectPolys('au'))
}
