% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_poly.r
\docType{data}
\name{ss_poly}
\alias{ss_poly}
\title{Utah's site-specific standard polygon shapes}
\format{An sf type polygon shapefile}
\usage{
ss_poly
}
\description{
Polygons containing site-specific standard designations and information.
}
\keyword{datasets}
