% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul_trophic.r
\docType{data}
\name{ul_trophic}
\alias{ul_trophic}
\title{Utah Lake trophic data}
\format{A data.frame with 729 rows and 15 columns}
\usage{
data(ul_trophic)
}
\description{
Utah Lake trophic data
}
\keyword{datasets}
