% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conv_table.r
\docType{data}
\name{unit_conv_table}
\alias{unit_conv_table}
\title{Unit conversion table}
\format{data frame}
\usage{
unit_conv_table
}
\description{
Table of common unit conversions and conversion factors for use in convertUnits function.
}
\keyword{datasets}
