% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut_poly.r
\docType{data}
\name{ut_poly}
\alias{ut_poly}
\title{Utah state boundary, excluding tribal jurisdictions (unofficial)}
\format{An sf type polygon shapefile}
\usage{
ut_poly
}
\description{
Polygon showing unofficial boundaries of the State of Utah, excluding tribal jurisdictions. This is an unofficial representation used to screen sites for the integrated report only.
}
\keyword{datasets}
