% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterYear.r
\name{waterYear}
\alias{waterYear}
\title{Assign water years to dates}
\usage{
waterYear(dates, start_month = 10)
}
\arguments{
\item{dates}{Input date vector. Must of class "Date".}

\item{start_month}{Month to start water year. Default is 10.}
}
\value{
Returns vector of water years.
}
\description{
This function assigns water years to standard dates.
}
\examples{
dates=as.Date(c("2022-09-29","2022-09-30","2022-10-01","2022-10-02","2022-10-03"))
waterYear(dates)
}
