% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmu_poly.r
\docType{data}
\name{wmu_poly}
\alias{wmu_poly}
\title{Utah DWQ watershed management unit polygons}
\format{An sf type polygon shapefile}
\usage{
wmu_poly
}
\description{
Utah DWQ watershed management unit polygons.
}
\keyword{datasets}
