% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqDD.r
\name{wqDD}
\alias{wqDD}
\title{Run the Water Quality Data Dashboard
Runs the Water Quality Data Dashboard app embedded in UT WQ tools package.}
\usage{
wqDD()
}
\description{
Run the Water Quality Data Dashboard
Runs the Water Quality Data Dashboard app embedded in UT WQ tools package.
}
