% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqpDD.r
\name{wqpDD}
\alias{wqpDD}
\title{Run the Water Quality Portal Data Dashboard
Runs the Water Quality Portal Data Dashboard app embedded in UT WQ tools package.}
\usage{
wqpDD()
}
\description{
Run the Water Quality Portal Data Dashboard
Runs the Water Quality Portal Data Dashboard app embedded in UT WQ tools package.
}
