% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMapResetButton.r
\name{addMapResetButton}
\alias{addMapResetButton}
\title{Add a map reset button to a Leaflet map object}
\usage{
addMapResetButton(map)
}
\arguments{
\item{map}{A Leaflet map object}
}
\description{
Add a map reset button to a Leaflet map object
}
\examples{
reset_map=wqTools::buildMap() \%>\% addMapResetButton()
}
