% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPolys.r
\name{assignPolys}
\alias{assignPolys}
\title{Assign spatial polygons to sites}
\usage{
assignPolys(
  polygon,
  columns,
  data,
  lat = "LatitudeMeasure",
  long = "LongitudeMeasure"
)
}
\arguments{
\item{polygon}{Input polygon used to assign information to sites. Must be an sf type polygon shapefile. User may choose from au_poly, bu_poly, huc8_12_poly, huc8_poly, huc12_poly,
ss_poly, ut_poly, or wmu_poly}

\item{columns}{Vector of columns user wishes to retain from polygon object. If columns object is not specified, function will retain all polygon columns.}

\item{data}{Input dataset. Must include latitude & longitude columns.}

\item{lat}{Name of latitude column. Default matches WQP objects.}

\item{long}{Name of longitude column. Default matches WQP objects.}
}
\value{
Returns the input data frame with assessment unit information appended.
}
\description{
This function assigns spatial polygon data to water quality portal type site objects (or data with site information attached).
}
